some improvements to message response and remanents of failures trying to setup application commands

+++ musicPlayer.py
@@ -11,7 +11,9 @@ import youtubeUtils
 class MusicPlayer(commands.Cog):
   """The music player command interface"""
 
-  def __init__(self):
+  def __init__(self, bot : commands.Cog):
+    self.bot = bot
+
     self.playing_video_id = None
     self.voice_channel = None
 
@@ -22,78 +24,90 @@ class MusicPlayer(commands.Cog):
     #   video_title = ...
     # }
 
+  # @commands.Cog.listener() # no me sale lo de los app commands
+  # async def on_ready(self):
+  #   print("MusicPlayer cog loaded")
+
+  # @commands.command()
+  # async def sync(self, ctx):
+  #   fmt = await ctx.bot.tree.sync(guild=ctx.guild)
+  #   await ctx.send(f"Synced {len(fmt)} commands")
+
+  # @discord.app_commands.command()
+  # async def test(self, interaction: discord.Interaction):
+  #   await interaction.response.send_message(content="wololo")
+
   @tasks.loop(seconds=1)
   async def sound_stream(self):
-    if self.queue and self.voice_channel:
-      video_id = self.queue.pop(0)["video_id"]
-      self.playing_video_id = video_id
+    try:
 
-      # buscamos si existe en el buffer, si no, intentamos descargarlo
-      if not os.path.exists(f"./buffer/{video_id}.mp3"):
-        await self.__download(f"https://www.youtube.com/watch?v={video_id}", video_id)
+      if self.queue and self.voice_channel:
+        video_id = self.queue.pop(0)["video_id"]
+        self.playing_video_id = video_id
 
-      # reproducimos en discord
-      if os.path.exists(f"./buffer/{video_id}.mp3"):
-        self.voice_channel.play(discord.FFmpegPCMAudio(f"./buffer/{video_id}.mp3"))
+        # buscamos si existe en el buffer, si no, intentamos descargarlo
+        if not os.path.exists(f"./buffer/{video_id}.mp3"):
+          await self.__download(f"https://www.youtube.com/watch?v={video_id}", video_id)
 
-        while True: # espera a que termine la canción 
-          await asyncio.sleep(2)
-          if not self.voice_channel.is_playing() and not self.voice_channel.is_paused():
-            break
+        # reproducimos en discord
+        if os.path.exists(f"./buffer/{video_id}.mp3"):
+          self.voice_channel.play(discord.FFmpegPCMAudio(f"./buffer/{video_id}.mp3"))
 
-        os.remove(f"./buffer/{video_id}.mp3")
+          while True: # espera a que termine la canción 
+            await asyncio.sleep(2)
+            if not self.voice_channel.is_playing() and not self.voice_channel.is_paused():
+              break
 
-    else:
-      await self.__kill_sound_stream()
+          os.remove(f"./buffer/{video_id}.mp3")
 
-  # def play_decorator(self, func):
-  #   async def wrapper(ctx, *args, **kwargs):
+      else:
+        await self.__kill_sound_stream()
 
-  #     func()
-
-  #   return wrapper
+    except discord.ClientException: #se levanta en caso de que este desconectado el canal, fallas en el internet probablemente...
+      print("voice client disconected!!!")
+      # self.voice_channel.connect()
+    except Exception as error:
+      print(error) 
 
   @commands.group(brief="play some music", description="reproduce the given youtube video or playlist url audio on voice channel", invoke_without_command=True)
-  # @play_decorator
   async def play(self, ctx, url = commands.parameter(default="", description="a youtube video or playlist url")):
-    if not ctx.author.voice.channel:
-      await ctx.send("conectate a un canal de voz")
-      return
+    if not ctx.author.voice:
+        return await ctx.send("conectate a un canal de voz")
     
     # conectamos a un canal, si no estamos en uno ya
     voice_channel = ctx.guild.voice_client
-    if not ctx.guild.voice_client:
+    if not voice_channel:
       voice_channel = await ctx.author.voice.channel.connect()
 
     self.voice_channel = voice_channel
-
+    
     # agregamos la(s) canciones a la queue
     videos_info = youtubeUtils.get_videos_info_from_url(url)
     if(videos_info):
       self.queue = self.queue + videos_info
     else:
       return await ctx.send("not a valid url") 
-
+    
     # elimina el mensaje para no causar spam
-    message_id = ctx.message.id
-    msg = await ctx.channel.fetch_message(message_id)
-    await msg.delete()
+    await ctx.message.delete()
 
     # inicia la reporduccion
+    more_text = f" and {len(videos_info) - 1} more" if len(videos_info) > 1 else ""
     if not self.sound_stream.is_running():
+      await ctx.send(f"playing \"{videos_info[0]["video_title"]}\"{more_text} ...")
       self.sound_stream.start()
     else:
-      await ctx.send("appending to queue...")
+      await ctx.send(f"appending \"{videos_info[0]["video_title"]}\"{more_text} to queue...")
 
   @play.command(name="search")
   async def play_search(self, ctx, query = commands.parameter(default="", description="a youtube video title, if using spaces this should be contain within \"\"")):
-    if not ctx.author.voice.channel:
-      await ctx.send("conectate a un canal de voz")
-      return
+    if not ctx.author.voice:
+        await ctx.send("conectate a un canal de voz")
+        return
     
     # conectamos a un canal, si no estamos en uno ya
     voice_channel = ctx.guild.voice_client
-    if not ctx.guild.voice_client:
+    if not voice_channel:
       voice_channel = await ctx.author.voice.channel.connect()
 
     self.voice_channel = voice_channel
@@ -109,9 +123,7 @@ class MusicPlayer(commands.Cog):
     })
 
     # elimina el mensaje para no causar spam
-    message_id = ctx.message.id
-    msg = await ctx.channel.fetch_message(message_id)
-    await msg.delete()
+    await ctx.message.delete()
 
     # inicia la reporduccion
     if not self.sound_stream.is_running():
@@ -125,17 +137,22 @@ class MusicPlayer(commands.Cog):
     if self.voice_channel:
       await self.__kill_sound_stream()
 
-  @commands.command(brief="skips to the next song on queue", description="stops audio playing for the current song (if any) and plays the next on queue")
-  async def skip(self, ctx):
-    self.voice_channel.stop()
+  @commands.command(brief="skip the current song", description="stops audio playing for the current song and plays another from the queue, by default the next one")
+  async def skip(self, ctx, queue_index = commands.parameter(default=0, description="skip to this position in the queue")):
+    if len(self.queue) - 1 >= queue_index:
+      await self.voice_channel.stop()
 
-    await ctx.send("skiping...")
+      await ctx.send("skiping...")
 
-    if os.path.exists(f"./buffer/{self.playing_video_id}.mp3"):
-      os.remove(f"./buffer/{self.playing_video_id}.mp3")
+      if os.path.exists(f"./buffer/{self.playing_video_id}.mp3"):
+        os.remove(f"./buffer/{self.playing_video_id}.mp3")
+      
+      # pop every song that is before
+      for _ in range(queue_index):
+        self.queue.pop(0)
 
-    if self.sound_stream.is_running():
-      self.sound_stream.restart()
+      if self.sound_stream.is_running():
+        self.sound_stream.restart()
 
   @commands.command(brief="pause playing", description="pause the current audio playing, to resume use !resume")
   async def pause(self, ctx):
@@ -207,10 +224,13 @@ class MusicPlayer(commands.Cog):
   async def shuffle(self, ctx):
     if self.queue:
       random.shuffle(self.queue)
+
       video_id = self.queue[0]["video_id"]
       info = youtubeUtils.get_video_snippet_from_video_id(video_id)
 
-      await ctx.send(f"queue shuffled, up next: {info["title"]}")
+      em = discord.Embed(title="queue shuffled", description=f"up next: {info["title"]}")
+
+      await ctx.send(embed=em)
 
   ########################   utils   ########################
   async def __kill_sound_stream(self):
@@ -237,3 +257,7 @@ class MusicPlayer(commands.Cog):
 
     with yt_dlp.YoutubeDL(options) as dl:
       dl.download([url]) #ver si aqui se puede evitar descargarlo y que se guarde en la ram
+
+
+# async def setup(bot: commands.Bot):
+#   await bot.add_cog(MusicPlayer(bot))
